Front end logic implemented
Added JavaScript logic to retrieve posts from the server currently with stubs where it is supposed to connect to the backend. Added a check to make sure user is part of the board before they are even able to do anything. If they're not, they get a special message and an option to request to join, which currently just denies them for now.

I also made minimal CSS changes to make the welcome text a little more readable by overlaying a gradient over the image. This is intended as a temporary fix, so please feel free to replace this with whatever styles you want, as long as it's aesthetically pleasing and the text is readable.