Implemented adding messages
Added ability to send messages. Upon sending a message, that message is
then opened. In the menu to make a post, you can only select channels
that you are a part of. The viewing message functionality ensures that
only members part of a channel can view the message.

Also, sorry for being a bit late with this feature, I'm gonna keep
working on this throughout the night and hopefully merge in the morning.