Visual changes
Added filtering by channel to fulfil requirement "If the user selects a board, then they enter a Slack-like (or Discord-like) interface where they can see the users and channels in the discussion board". The part with users is left as an optional requirement in the storyboard, but I added a channel selector just so the people marking can see the channels.