Created README.txt
Please add your resposibilities to the README.txt file. The TA's and the
prof will calculate the mark based on the sections that you indicate.

Also I added an index.php to the public_html folder. The idea is that if
the user puts in the link https://cs.mcgill.ca/~mlavre1/ , they will be
redirected to https://cs.mcgill.ca/~mlavre1/landingpage.html and not see
the index of our public_html folder.