Fixed course applications
You can now apply to courses, and it will show an error if you are already
part of the course, or have already applied to the course. Otherwise you
should be successfully added to the applicants list. Also fixed the
"Back" link in each of the php pages for managing.